feat(languages): adding dynamic language support (#6)
### Related Ticket(s)

Refs #4 #5 

### Description

This PR updates so that the user no longer needs to configure any language support in the webpack configuration. Pages and language folders should be now autogenerated based on the locale data. Language/locale data is also now dynamically generated in the base page as well.

If a user only has one language folder, the folders won't be generated and will instead generate the pages in the root of the distribution folder. The locale button will also be automatically disabled.

### Changelog

**New**

- Automated language support

**Changed**

- webpack config, language files
- added test locale folder